Change ZuluPhysicalProperties to not be abstract and update the default values to remove the HumanoidURDFParameterInterface.java and ZuluPhysicalPropertiesV0.java.
