+++ zulu/src/main/java/us/ihmc/zulu/ZuluModelFactory.java
@@ -1,7 +1,6 @@
 package us.ihmc.zulu;
 
 import org.apache.commons.lang3.SystemUtils;
-import us.ihmc.zulu.parameters.model.HumanoidURDFParameterInterface;
 import us.ihmc.euclid.transform.RigidBodyTransform;
 import us.ihmc.modelFileLoaders.RobotDefinitionLoader;
 import us.ihmc.multicastLogDataProtocol.modelLoaders.DefaultLogModelProvider;
@@ -13,6 +12,7 @@ import us.ihmc.scs2.definition.robot.WrenchSensorDefinition;
 import us.ihmc.scs2.definition.robot.urdf.URDFTools;
 import us.ihmc.scs2.definition.robot.urdf.items.URDFModel;
 import us.ihmc.wholeBodyController.RobotContactPointParameters;
+import us.ihmc.zulu.parameters.model.ZULUURDFParameters;
 
 import java.util.Arrays;
 import java.util.function.Consumer;
@@ -21,7 +21,7 @@ import java.util.function.Predicate;
 public class ZuluModelFactory
 {
    private final String[] resourceModelsToBeLogged;
-   private final HumanoidURDFParameterInterface urdfParameters;
+   private final ZULUURDFParameters urdfParameters;
    private final RobotContactPointParameters<RobotSide> contactPointParameters;
    private RobotDefinition simulationRobotDefinition;
    private final ZuluVersion zuluVersion;
