--- zulu/src/main/java/us/ihmc/zulu/parameters/model/HumanoidURDFParameterInterface.java
@@ -1,15 +1,0 @@
-package us.ihmc.zulu.parameters.model;
-
-import java.io.InputStream;
-
-public interface HumanoidURDFParameterInterface
-{
-
-   String getURDFModelName();
-
-   String[] getResourceDirectories();
-
-   String[] getLoggedResources();
-
-   InputStream getURDFAsInputStream();
-}
