+++ zulu/src/main/java/us/ihmc/zulu/parameters/model/ZULUURDFParameters.java
@@ -16,7 +16,7 @@ import java.util.List;
  * This class contains the paths to the individual URDF files that will be used to build the robot. It also holds some simple methods to combine the different
  * files into one input stream that will be used to build the URDF.
  */
-public class ZULUURDFParameters implements HumanoidURDFParameterInterface
+public class ZULUURDFParameters
 {
    public static final String URDF_MODEL_NAME = "Zulu";
    private static final String[] RESOURCE_DIRECTORIES = new String[] {"zulu/", "zulu/urdf/", "zulu/meshes/"};
@@ -25,7 +25,6 @@ public class ZULUURDFParameters implements HumanoidURDFParameterInterface
 
    public static final String URDF_FULL_BODY = "zulu-full-body.urdf";
 
-   private final String[] robotModelResourceDirectory = new String[1];
    private final String[] urdfResourceDirectories;
    private final String[] urdfResourcesWithPath;
 
@@ -34,9 +33,6 @@ public class ZULUURDFParameters implements HumanoidURDFParameterInterface
    public ZULUURDFParameters(ZuluVersion zuluVersion)
    {
       urdfParserProperties.setHandleImplicitJointDefinitions(false);
-
-      // Directory containing all robot model resources for the given robot version
-      robotModelResourceDirectory[0] = zuluVersion.getRobotModelResourceDirectory();
 
       // List of directories containing robot model URDF resources for the given robot version
       List<String> urdfResourceDirectoriesList = new ArrayList<>();
@@ -61,25 +57,21 @@ public class ZULUURDFParameters implements HumanoidURDFParameterInterface
       urdfResourcesWithPath = urdfResourcesWithPathList.toArray(new String[0]);
    }
 
-   @Override
    public String getURDFModelName()
    {
       return URDF_MODEL_NAME;
    }
 
-   @Override
    public String[] getResourceDirectories()
    {
       return RESOURCE_DIRECTORIES;
    }
 
-   @Override
    public String[] getLoggedResources()
    {
       return LOGGED_RESOURCES;
    }
 
-   @Override
    public InputStream getURDFAsInputStream()
    {
       List<InputStream> inputStreamList = new ArrayList<>();
