+++ zulu/src/test/java/us/ihmc/zulu/ZuluLiftOffAndTouchDownTest.java
@@ -2,7 +2,7 @@ package us.ihmc.zulu;
 
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
-import us.ihmc.zulu.parameters.model.ZuluPhysicalPropertiesV0;
+import us.ihmc.zulu.parameters.model.ZuluPhysicalProperties;
 import us.ihmc.avatar.AvatarLiftOffAndTouchDownTest;
 import us.ihmc.avatar.drcRobot.DRCRobotModel;
 import us.ihmc.avatar.testTools.scs2.SCS2AvatarTestingSimulation;
@@ -14,27 +14,13 @@ import static org.junit.jupiter.api.Assertions.*;
 
 public class ZuluLiftOffAndTouchDownTest
 {
-   private final double footLength = new ZuluPhysicalPropertiesV0().getFootLengthForControl();
+   private final double footLength = new ZuluPhysicalProperties().getFootLengthForControl();
 
    @Tag("humanoid-flat-ground")
    @Test
    public void testForwardStepRotated() throws SimulationExceededMaximumTimeException
    {
       DRCRobotModel robotModel = new ZuluRobotModel(ZuluVersion.V1_FULL_ROBOT)
-      {
-//         @Override
-//         public WalkingControllerParameters getWalkingControllerParameters()
-//         {
-//            return new ZuluWalkingControllerParameters(getRobotVersion(), getTarget(), getJointMap(), getPhysicalProperties(), getContactPointParameters())
-//            {
-//               @Override
-//               public double nominalHeightAboveAnkle()
-//               {
-//                  return 0.76;
-//               }
-//            };
-//         }
-      };
       SCS2AvatarTestingSimulation testHelper = AvatarLiftOffAndTouchDownTest.setupTest(robotModel, Math.toRadians(90.0));
       // We need to turn off the CoP threshold fraction here. If it's left on, contact isn't triggered until the CoP is far enough within the foot. This means
       // that the detection of contact is significantly delayed. It then misses the cotnact timing, and also ahs a tendency to "slap" the foot onto the ground,
